Issue 472339: DevTools: provide affordance during recording
Issue 472339: DevTools: provide affordance during recording.
Added a css class called ‘pulse’ which has a css animation keyframe
associated with it called ‘pulse-effect’. The class is being added to
the recordButton in NetworkPanel.js in order to keep the pulse effect
limited to only the record button in Network Tab.