Devtools Animations: Don't wait for empty transitions to start
Some animations can be triggered but never started, such as transitions
which are synchronously triggered and cancelled in the same frame. As these
animations are never started, the front-end needs to be notified to not wait
for these animations to start.

BUG=447083

Review URL: https://codereview.chromium.org/1433623002

Cr-Original-Commit-Position: refs/heads/master@{#358739}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 17c1e863ea421cc8583a749120edba9e35e98a31