DevTools: fix css pretty-print.
The CSS pretty-printing was regressed back in the days when all formatters
were moved to a single content builder.

The content builder assumes that added tokens are "words" - and makes sure
those do not stick together. However, CSS formatter violates this contract.

This patch is a workaround which could be merged.

BUG=614678
R=dgozman

Review-Url: https://codereview.chromium.org/2012953002
Cr-Original-Commit-Position: refs/heads/master@{#396332}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: aec0b6fa92b2f2bece418aede111a78b043a0655