DevTools: cleanup treeoutline expand logic
TreeElement.expand() should not need to check that the element has children and
TreeOutline.insertChild() should not automatically expand a child when it is
inserted. ObjectPropertiesSection.js depended on this logic to automatically
expand the Entries property. This CL updates it so that the Entries property is
added before expanded.

BUG=

Review-Url: https://codereview.chromium.org/2147343002
Cr-Original-Commit-Position: refs/heads/master@{#407061}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: cf5b729f072ba670c7b9966407fb3d47eb2169e7