DevTools: Find bookmarks at line endpoints when calling _clearBookmarks
This fixes an issue with the source color picker where undo-ing the color text did not
always remove the color swatch. CodeMirror findMarks() is not inclusive, so we also need
to check the line endpoints.

The additional newline added to the string in _updateColorSwatches allows
color text to be matched when it is located at the start of a line.

R=lushnikov

Review-Url: https://codereview.chromium.org/2187173002
Cr-Original-Commit-Position: refs/heads/master@{#408758}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a267f68a32a1ecc8f1c220f621db70c96fe3738c