DevTools: [CMTE] hide autocomplete when cursor jumps over smart brace
If the cursor is located right before closing brace and user types in
")", then cursor just jumps over the brace. This is a "smart brace"
functionality.

In this case, it is important to close autocomplete. Today this doesn't
happen because the onChange event is not sent in this case.

BUG=none
R=einbinder

Review-Url: https://codereview.chromium.org/2382323002
Cr-Original-Commit-Position: refs/heads/master@{#422214}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d44da28cca2db51a7cb2132bed7bff26e8e96bef