DevTools: teach network UISourceCodes to return metadata
The patch teaches network-based UISourceCodes to return non-null
metadata.

For the resource-based UISourceCodes, the metadata simply comes
from the underlying resource.

For the sourcemap-originating UISourceCodes which content was
embedded in the sourcemap, the contentSize metadata is available.

Drive-By: improve complexity of WI.ResourceTreeModel.resourceForURL
from O(|resources|) to O(|frames|).

BUG=none

Review-Url: https://codereview.chromium.org/2408203007
Cr-Original-Commit-Position: refs/heads/master@{#425540}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: a22a11e3879975ab54daa514193934e4b32a966e