DevTools: properly restore selected DOMNode in Elements panel.
Since the http://crrev.com/b4d6bf98e, we started
sending two "documentUpdated" events for every
page reload.

It's not a bad thing by itself (but probably needs
addressing in a follow-up), but the element restoration
logic is unprepared for this.

This patch re-writes the element restoring logic in the
elements panel.

BUG=645645

Review-Url: https://chromiumcodereview.appspot.com/2428823002
Cr-Original-Commit-Position: refs/heads/master@{#426525}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9cb88eeb6589fb10e8b7e87c76b54ee16d2f844e