DevTools: [Persistence] tabs should not change order during reload
During page reload, the network UISourceCodes are removed and then
re-added. This results in bindings being removed and then re-added.

However, currently we're not doing a good job maintaining the selected
tab and the tab order - they're getting shuffled during this process.

This patch fixes the behavior so that the reload does not mess tabs
order.

BUG=660596
R=dgozman

Review-Url: https://codereview.chromium.org/2460513003
Cr-Original-Commit-Position: refs/heads/master@{#428555}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 1a31eac72f8252705bac3cee5823d0ec92d8d857