[DevTools] refactor response to arrow key presses into TreeElement
This creates the methods "descendOrExpand" and "collapseOrAscend" to TreeElement, and calls those methods in TreeOutline._treeKeyDown. This is to abstract the existing behaviour out of the _treeKeyDown method so that it may be overridden (and it also breaks up the logic of _treeKeyDown a bit).

BUG=560525

Review-Url: https://codereview.chromium.org/2469793002
Cr-Original-Commit-Position: refs/heads/master@{#429671}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: ebc67fff2116008879e438020f0b27edec80efb9