[DevTools] added BreakpointManager.possibleBreakpoints method
This method takes UISourceCode and TextRange and returns UILocations of possible breakpoints in this range.
It uses first target with JS capability that was able to resolve startUILocation and endUILocation to call getPossibleBreakpoints on.
JavaScriptSourceFrame uses this method to get more accurate breakpoint location.

BUG=chromium:566801
R=lushnikov@chromium.org,dgozman@chromium.org

Review-Url: https://codereview.chromium.org/2484283004
Cr-Original-Commit-Position: refs/heads/master@{#431350}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b363145ee1c6ca60e5dfa882d78b3430e3d680ca