DevTools: extract checkbox checkmark from toolbarButtonGlyphs
This patch extracts the checkbox checkmark into a separate image file.
The checkmark is used in the themed colorful checkboxes. One example
of such a checkbox is checkboxes "Loading" and "Scripting" in the
EventLog of timeline panel.

Unfortunately, this can't be migrated over to UI.Icon because checkboxes
are not allowed to have descendants.

BUG=669323
R=dgozman_slow

Review-Url: https://codereview.chromium.org/2663383002
Cr-Original-Commit-Position: refs/heads/master@{#447653}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: f241c785086352667c80a95a20362fe88bd19f02