DevTools: restore the style of expandable titles on console object previews
This CL reverts changes to the rendering of console objects. The goal is to
restore clarity to the fact that text shown to the right of the arrow is a stale
snapshot while items in the tree expansion are computed live once on expand.

Expanded titles will just show the 'i' note, instead of an abbreviated title:
https://codereview.chromium.org/2139043002

Italic titles will be restored, once removed in:
https://codereview.chromium.org/2514123003

BUG=688137

Review-Url: https://codereview.chromium.org/2676773002
Cr-Original-Commit-Position: refs/heads/master@{#447934}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 9d8e815220afc20f198f0b513cb46edc2d823186