DevTools: fix stretch layout in search/replace bar
SearchableView used to build the find and replace UI using <table>. A table
inside a flex container produced the strange behavior in the bug.

This CL fixes the visual bug, removes the table elements, and changes the layout
of the find/replace text inputs to stretch as much as possible.

BUG=698830

Review-Url: https://codereview.chromium.org/2734043002
Cr-Original-Commit-Position: refs/heads/master@{#455244}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 336d5878c4c612627f0a8333f71b50a04b7091d1