[Devtools] Network waterfall to use Path2D to draw in layers
This patch introduces changes network waterfall canvas to use Path2D to
draw it's canvas items. This is in preporation to move it to use full
layers. (Followup patch will move the code to use the same path
for artifacts that are the same style)

R=caseq
BUG=697668

Review-Url: https://codereview.chromium.org/2731633002
Cr-Original-Commit-Position: refs/heads/master@{#455992}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 09a524bf16691a0fd873d1d24edf04c8aae8874d