[Devtools] Fix offline checkbox state sharing across Network and Application>Service Workers panels
Fixes an issue where the last network condition prior to going offline was not shared between offline checkboxes (NetworkConditionsSelectors).

Due to this change, it is also no longer necessary for each NetworkConditionsSelector to update the state in the networkConditionsChanged() callback and so this has been removed.

BUG=702070

Review-Url: https://codereview.chromium.org/2757093002
Cr-Original-Commit-Position: refs/heads/master@{#458699}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 8e2177f8dde7e53ab585a1fb09dc3e69dd81c78b