DevTools: split text_utils out of common module
This patch introduces a new 'text_utils' module which has text-related
classes from 'common' module.

We always had a need to have text_utils in formatter worker, but since it
was part of a big 'common' module, re-implemented parts of it functionality
inside worker.

This is a pre-step to switching javascript outline from tokenizer to AST.
Generating AST with locations for every node is time-consuming; instead, we'd
like to simply re-use TextUtils.TextCursor.

BUG=438203
R=dgozman

Review-Url: https://codereview.chromium.org/2769843003
Cr-Original-Commit-Position: refs/heads/master@{#459357}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da678dd14198b5dee5303e9d51d1ff883b4a8e12