DevTools: prepare tests for Persistence2.0 release
We have a bunch of tests which rely on DefaultMapping to test
Persistence layer. However, they don't test the DefaultMapping
behavior - instead, they want to test the Persistence layer and everyone
who depends on it.

This patch:
- adds a simple TestMapping which controllably adds and removes persistence
  bindings
- moves majority of the tests over to TestMapping
- removes a few tests which were testing themselves

This is a pre-step to enable Persistence2.0.

BUG=277885
R=dgozman

Review-Url: https://codereview.chromium.org/2784733002
Cr-Original-Commit-Position: refs/heads/master@{#460466}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6b74f6642be4ca9d5af4de7dcc56ea58c3f62e53