[DevTools] Remove SDK.SDKObject
We have a few users of it, which are better to use underlying models.
We also have a log of SDKObject's like classes which do not inherit from SDKObject -
that means this abstraction doesn't really work.

As a bonus, we only inherit from Common.Object when necessary, reducing the overhead.

BUG=none

Review-Url: https://codereview.chromium.org/2801943003
Cr-Original-Commit-Position: refs/heads/master@{#462949}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 91588b92f697f08d822778db08d05886a23c51fe