[Devtools] New structure and colorize rows for network products
Refactored the way the dependency chain is going for the products
module. We now have a shared module that is bundled and can be used to
resolve things like frame and product for url.

We are also adding background coloring to network rows.

R=pfeldman
BUG=712876

Review-Url: https://codereview.chromium.org/2839273003
Cr-Original-Commit-Position: refs/heads/master@{#468203}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 56637492051aec80a5023fc0bd8314c0143241db