DevTools: only show one source map infobar at a time
The 'Source map detected' infobar is created when a script with a sourcemap is
updated. When Persistence is enabled, this may happen each time the user
refreshes the page, resulting in many duplicate infobars.

This CL checks for an existing sourceMapInfobar before creating a new one, or
removes it if the updated script has no sourcemap.

BUG=716340

Review-Url: https://codereview.chromium.org/2851483003
Cr-Original-Commit-Position: refs/heads/master@{#468509}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 15d1662e1a70f83a71e629aa7daa43c3d887dcd0