[Devtools] HAR format now uses issueTime & adds chrome specific timings
Move HAR format to use issueTime rather than startTime for HAR's
startTime. This was an because HAR is not friendly to negative timings
(negative often means there is no data) and when importing HAR we need
to know when the time between when the issueTime (when the browser found
out about a request) and the startTime (when the request was
dispatched); we show this result as "Queued". The delta between when the
request was dispatched (startTime) and when the min(dnsStart,
connectStart, sendStart) is the "Blocking" or "Stalled" time. We were
showing only Blocked/Stalled for blocked because we we would be using
negative numbers to represent the "queued" time.

R=caseq
BUG=374557

Change-Id: I157518cdcf67009e26c81fee96c239d612eee28a
Reviewed-on: https://chromium-review.googlesource.com/558504
Commit-Queue: Blaise Bruer <allada@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#487351}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d225d794a786d739206f24fb9d48787bedf296c0