DevTools: generalize filter suggestion builder part 1
The goal of this CL is to remove network-specific sorting logic from
FilterSuggestionBuilder, in order to allow console to use it.

Part 2 is simply to move to a new location. This makes it easier
to review part 1.

Bug: 685283
Change-Id: I50d5632e97a646f4204016a5eae41d3cfe7b17a2
Reviewed-on: https://chromium-review.googlesource.com/669762
Commit-Queue: Erik Luo <luoe@chromium.org>
Reviewed-by: Alexei Filippov <alph@chromium.org>
Reviewed-by: Blaise Bruer <allada@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#508823}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: db58678d52c048df084359b377adb1e0dffe1eb5