DevTools: deflake application panel tests by resetting storage state
This creates an ApplicationTestRunner.resetState() method to reset storage state
and is added to the beginning of every test that relies on a storage API.

This makes ClearDataForOrigin more robust by making it explicit that it's
asynchronous and (mostly) fixing the race condition.

The callback doesn't guarantee the storage has actually been deleted but
at least it has been scheduled (see storage_partition.h), whereas before
it returned with a success response immediately.

We could update third_party/WebKit/Source/devtools/front_end/resources/ClearStorageView.js
to actually wait for this callback to be called to update the UI when the user
clicks on the clear storage button rather than rely on a 500ms timeout,
but let's defer this for a follow-up patch.

Bug: 667560
Change-Id: Ia7b70aebabd970957ae3581e47ee31a10992c657
Reviewed-on: https://chromium-review.googlesource.com/792534
Commit-Queue: Will Chen <chenwilliam@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#524555}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0062d697524210287d8da8758e3cd9c9689aacd1