Use host timestamp for devtools screencast
During remote debugging, the host is a more reliable source of timestamp
than the remote debugging agent. This CL makes the timestamp optional in
for EmulateTouchFromMouseEvent, and stops supplying the timestamp from
input methods on the frontend. The debugging agent retains the option to
provide a custom timestamp if necessary.

No test is included for the frontend fix. A test is added for the
backend to check empty timestamp is accepted for
EmulateTouchFromMouseEvent, and that the default timestamp comes from a
monotonic clock.

Bug: 797334
Change-Id: I2f01cf225cd2db6a1315f7c2c2a9305a0c522643
Reviewed-on: https://chromium-review.googlesource.com/850728
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Deepanjan Roy <dproy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#530957}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5ba3218966ea7d7063bbc00577778952b53c463a