DevTools: support latest sourcemaps for script hot reload
There are cases when multilpe sourcemaps refer to the same sources:
1. this might happen when bundles share common code
2. this is also the case with hot module reload: every reload
   injects a new script with a new sourcemap that refers to the same
   sources.

		+--------+
      +-------> | foo.ts |
      +         +--------+
     SM1
      +
      +-------> +--------+
		| bar.ts | <-- overlapping source file
      +-------> +--------+
      +
     SM2
      +         +--------+
      +-------> | baz.ts |
		+--------+

In the first case, contents of overlapping sources are equal and thus don't
deal any trouble.

In the second case, contents of overlapping sources are different; the latest
injected sourcemap has the most up-to-date version of the content.

This patch teaches CompilerScriptMapping to re-create UISourceCodes for the
sourcemap sources if a new SourceMap got attached referring to the same source.
This behavior satisfies both cases.

BUG=438251
R=dgozman, kozy

Change-Id: I98384eed39d6b347d466538fc9151a0c7b798a95
Reviewed-on: https://chromium-review.googlesource.com/875419
Commit-Queue: Andrey Lushnikov <lushnikov@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Aleksey Kozyatinskiy <kozyatinskiy@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#531004}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 26814c4b59a440949b0a0a588f3c2cfd0d391bc3