DevTools: Recycle SourceFrames on persistence binding
Previously, when a network UISourceCode was open, and a persistence
binding appeared, the user was transferred over to a new SourceFrame
for the filesystem UISourceCode. This resulted in an annoying flicker,
and caused a loss of undo history.

Bug: 778043
Change-Id: Ie846f96d3c4f2553e8a78db5259f394e358f94fa
Reviewed-on: https://chromium-review.googlesource.com/1015195
Commit-Queue: Joel Einbinder <einbinder@chromium.org>
Reviewed-by: Andrey Lushnikov <lushnikov@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#552267}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4435c448310d3850f31c65509f2dfce32ddf594e