DevTools: display colors in suggest box for CSS variables
This patch starts displaying color swatches in suggest box
next to css variables that compute to color value.

For this to be possible, the ColorSwatch and BezierSwatch were
moved under the ui/ module. The CSSShadowSwatch stayed in the
inline_editor/ for now because it has a dependency on
ShadowModel.

R=einbinder, dgozman

Change-Id: Ic29a13073dc8236d5ef9a2fe35adf6bcd927290b
Reviewed-on: https://chromium-review.googlesource.com/1026882
Commit-Queue: Andrey Lushnikov <lushnikov@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#553468}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: b85cc4539f06a76af94addf2d193adaa4edc9b9d