DevTools: do not show tooltip for hidden elements
In StylesSidebarPane, the 3-dot 'More tools' icon reveals a toolbar on hover.
On 'mouseenter', we set 'display: none' on it, so the following 'mousemove'
determines its bounding rect as {x: 0, y: 0, w: 0, h: 0}, and shows a tooltip
in the top-left of DevTools.

This CL removes the unnecessary tooltip and checks that an element is visible
before showing the tooltip.

Bug: 829342
Change-Id: Ib0268f70dccbb1190916796e1ca6aadaec763e5f
Reviewed-on: https://chromium-review.googlesource.com/1067784
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Erik Luo <luoe@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#560738}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 67c89d4caa1ffd0b7d30d490e68ae9bd2dc1c544