Add a link to view certificate in DevTool's SignedExchangeInfoView.
It is useful when debugging the certificate especially when failed to
load the signed exchange.
(When succeeded, we can view the certificate in Security panel)

This CL also changes "certUrl" to optional in PDL file, because this
could be null according to the spec.

Bug: 830505
Change-Id: I710d91bc43bd414d03aca7d4e3dfdb9342b10500
Reviewed-on: https://chromium-review.googlesource.com/1055407
Commit-Queue: Tsuyoshi Horo <horo@chromium.org>
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Reviewed-by: Kunihiko Sakamoto <ksakamoto@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#561758}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4defbb6af0fda52f8a5ccc32ad5eb03cb31ce645