[DevTools] Prevent SW-from-other-domain's list from collapsing on search
Add an event.target check in SW-from-other-domain's filter, so that an
Enter or Space key input in the filter's editor box will not collapse
the list of service workers from other domains. However, if the focus is
on the div element of the "service worker from other domains" (instead
of on the editor box inside the div), an Enter or Space key is still to
expand/collapse the list.

Bug: 855001
Change-Id: Iaeedd47fefe64bb081cd7cae66fd3eb53b4418b0
Reviewed-on: https://chromium-review.googlesource.com/c/1315356
Reviewed-by: Erik Luo <luoe@chromium.org>
Commit-Queue: Haihong Li (Harley) <hhli@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#605080}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 5fc2dfd4633fc10b771c3efd53eef8133cd0d282