[DevTools] Prevent deleting text from deleting entire entry in IndexedDB
Use UI.isEditing() is DataGrid._keydown() event handler to check if the
DataGrid is being edited.

When a delete key event arrives, event.target is always the div element
(the element that surrounds the textarea being edited) so one needs to
resort to document.deepActiveElement() defined in DOMExtension.js, to
check if a textarea is being edited, and UI.isEditing() is a wrapper
for it. If UI.isEditing() returns true, the entry that textarea belongs
to should not be deleted.

Bug: 899545
Change-Id: I598677d2cdccf90df5f8177be68a1ef1d4642938
Reviewed-on: https://chromium-review.googlesource.com/c/1383037
Reviewed-by: Erik Luo <luoe@chromium.org>
Commit-Queue: Haihong Li (Harley) <hhli@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#617689}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: c4f5cc233f0d79a679046af9245de69daeee4f3b