DevTools: Fix 4px of white space in network big rows mode
When the network is in large rows mode, the heading is larger. The
tabbed pane when a request is selected had a margin added to match the
larger heading. But that caused a strange white space to appear above
the tabbed pane.

Change-Id: I19bdf61509b2f14c80e06113228475759da9e7d2
Reviewed-on: https://chromium-review.googlesource.com/c/1387906
Reviewed-by: Dmitry Gozman <dgozman@chromium.org>
Commit-Queue: Joel Einbinder <einbinder@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#625690}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 6fffcad11f21e73856362519bec47e572e9f8450