DevTools: fix position of gutter decorations
Positions of gutter decorations regressed in [1] when an element
that used to be display: none became visibility: hidden.

This CL fixes the positions of decorations to be at the top of
the row.

Screenshot: https://imgur.com/a/RnUoRrT

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1702746

Bug: 994296
Change-Id: I8bccaa8c8e1b1f77d905c5b66fc43431f5733e4b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1769255
Reviewed-by: Jeff Fisher <jeffish@microsoft.com>
Commit-Queue: Erik Luo <luoe@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#691195}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 41418672e83ed391126c0ecae5df5119ad7665c1