[devtools] Don't show errors in Sources panel popovers.
When hovering an expression within the Sources panel (during debugging),
we try to evaluate it's value and show a popover with the result of
that. However this is all best effort and might fail for various
reasons. If it fails because we guessed the wrong substring (for an
expression), we currently display a popover with a SyntaxError or a
ReferenceError in that. But this popover is confusing and doesn't serve
a purpose. Instead we should just not display anything in this case as
well.

Bug: chromium:995036
Change-Id: Id70e2212fe0af58078b2a33731688ec1c5507434
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1831814
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#701465}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 4dc425332d3155428f831df39e11bf7f2c1ed842