Devtools: New "Initiator" tab in Network Item view
Issue: In the network tool, there is a pop up for stack traces that appears
when you hover over the initiator column. This info is not available
anywhere else, and is not keyboard accessible.

This changes: add a new tab to the network item view, similar to the
Timings tab and Waterfall column (which is available when hovering over the
waterfall, but also accessible when selecting a network request to view).

Bug: 963183, 842488

workflow gif: https://imgur.com/a/wmdNYXE

Change-Id: Ib2f23b924a06e687acd362ccf4d6764f8a996559
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1742990
Commit-Queue: Junyi Xiao <juxiao@microsoft.com>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Joey Arhar <jarhar@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#702125}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: 0d0cae574b458c3d4bd7f6f254f74f6f2442b710