Remove terminal/ and the corresponding experiment
The terminal experiment was initially proposed in 2016 [1]. Since then,
no maintenance work was performed on promoting the experiment to a
stable feature.

Closer inspection of the feature shows that it does not work. After
enabling the experiment, issueing the command "Show Terminal" results in
an internal error in xterm.

If we want to reconsider this experiment, we should re-evaluate the
approach and come up with a proper design for the feature. For now, I
would propose on removing the dead code.

[1]: https://codereview.chromium.org/2372303003

Bug: 1006759,1011466
Change-Id: Ib433b44924d9ccb1f5665afba9affcccd04c2bd8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1855923
Commit-Queue: Yang Guo <yangguo@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Cr-Original-Commit-Position: refs/heads/master@{#705165}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: da5f866faa7e93ca75874023017fb6ff8aa5b75b