Devtools: IndexedDB ContextMenu Expand and Collapse Value Objects
This change adds 2 context menu entries for the IndexedDB datagrid
to allow for expanding and collapsing the value objects from the keyboard.

Note: This change does not currently work with the contextMenuKey
due to an existing issue where contextMenuKey on
the data grid will not target the selected
node, but on whichever row happens to be in the center
of the currently visible section of data grid. This issue will be
resolved with this change:
https://chromium-review.googlesource.com/c/chromium/src/+/1729554


Before: https://imgur.com/8Iz96jv

After: https://imgur.com/UwSUVOh

Bug: 963183
Change-Id: Iefb563dc744b524d3282b62141a9da12230e9911
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/1725990
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>
Cr-Original-Commit-Position: refs/heads/master@{#706603}
Cr-Mirrored-From: https://chromium.googlesource.com/chromium/src
Cr-Mirrored-Commit: d0bf0d06e4239335f5cbe022724ac3ae29ea2563