DevTools: Elements renderer shouldn't raise errors
Occasionally, the DevTools frontend will attempt to render an
element which has gone away in between the initial notification
and a later render (such as an expansion to show that element).
In cases like these, the renderer will throw an exception (it will
actually return a rejected promise, but these are functionally
equivalent). However, the render function contract already does
not guarantee that it will return a concrete value, and the only
two call sites that care about it already check for that case.

Because these error cases are unobserved, these cases are ignored,
and so the code paths should be removed.

Change-Id: I9ad444b130ab80779daa41edd9fcf03815a6a720
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1872432
Commit-Queue: Lorne Mitchell <lomitch@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>