Harden Common.ResourceType.fromMimeType
In cases where a network request is blocked or cancelled,
the mimeType field remains 'undefined' because the
'responseReceived' event is never raised.

Common.ResourceType.fromMimeType requires that a valid
string is passed in, but there are instances where that
is not true. Instead of propagating logic like this:

  Common.ResourceType.fromMimeType(mimeType || '')

this change proposes that we harden the function to
handle this case.

Change-Id: Ieee1152e8a45aff24c90da7095413f816b83850c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1874335
Commit-Queue: Mike Jackson <mjackson@microsoft.com>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>