Continue stepping over in auto-stepping
When auto-stepping is enabled and user pressed "step over" instead of "step in", then continue repeating the same command so that we step over a function in the source code.

Before this change, devtools would step into a function and stop there on a resolved location regardless of the initial command.

Change-Id: I59444ccc0f5d6533070867cae433ff3e4b5e6815
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1881565
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Ingvar Stepanyan <rreverser@google.com>