[CSS Overview] Reduce the eagerness of reset
We currently observe the InspectedURLChanged event to decide if the
CSS Overview needs to be reset, but this can be fired, it seems, for
changes internal to the page (seems like iframes, perhaps). As such
with this CL we confirm that the inspected URL itself has changed
before triggering the reset.

Change-Id: I32b721b66bd75875ecd61426573131407c455a4f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1887791
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>