Devtools: Add a Request initiator chain section to Initiator tab in Network Item view
We currently have a similar functionality which using red/green to highlight
those requests in the network data grid when shift + hover. It's neither very easy to
discover nor accessible.
This change add a request initiator chain section to the initiator tab, which includes
a list of request URL that current request initiated and a list of request URL that
initiated current request in a tree structure.

screenshot: https://imgur.com/a/XCzczlA

Bug: 963183, 842488
Change-Id: If10071312b0cda36932014d4209afc17bab93ba3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1874670
Commit-Queue: Junyi Xiao <juxiao@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>