DevTools: Fix Debugger hover tips not working
A change was made recently to limit eager popover evaluation and limit popup availability to only when the variable was in scope. This change exists here: https://chromium-review.googlesource.com/c/chromium/src/%2B/1835357

However, in case of pretty-printed files and sourcemaps, the scope was getting calculated based on the transpiled JavaScript file, or in case of pretty-printed files, based on the original minified file, which is why the hover tips failed to show sometimes.

Fixing by checking if we can get any mapped locations for the scope start/end locations, if not, we fallback to the JS scope start/end locations.

Bug: 1015367
Change-Id: I6e91134242d5c3de349504b208dd93fb9366e8e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1928582
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Commit-Queue: Anubha Mathur <anubmat@microsoft.com>