Add script to remove unused globals
It Traverses all expression statements that put stuff on the global
scope and checks if it is only used once (e.g. in that particular
assignment). Anything that is only exported but never used is deleted.

This works for extensions declared in the module.json as well.

Change-Id: Ie3d9759da8593f8c66f1c12ba2e8ee1724f240d6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1942289
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>