Migrate test_runner/ to ESM
The TestRunner logic breaks when trying to load as an ES module. This
was the root cause for the failures we saw when we were migrating main/
to ESM.

Since TestRunner directly relies on main/ existence, we first have to
migrate the test_runner to ESM. However, this uncovered a lot of
problems. Most importantly, symbols were sometimes overwritten and
ordering was important.

The bootstrap methods of TestRunner are now moved to the top. This
includes setting up the error-catching logic and adding results.

All other symbols are exported but also put on the global scope, as that
is what the layout tests still require.

A pre-existing bug is the fact that tests time out if there is any
(syntax) issue in TestRunner itself. We will need to fix that issue, as
it makes debugging TestRunner itself very difficult. However, I will
punt that to a follow-up CL, as this CL is already quite difficult to
reason about.

Bug: 1006759, 1019182
Change-Id: Id64422968171888dbdbcc927dfd174d7868f333e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1953583
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>