DataGrid: Add aria-hidden to header table
This CL adds the aria-hidden tag to the header table in the datagrid.
Previously, hidden headers were added to the dataTable of datagrids
in order to give context to screen readers for the columns in a
datagrid. This caused screen readers to read out the headers twice,
once for the header table, and once for the data table.

This change hides the header table from screen readers, as the information
is already present in the data table.

Before: https://imgur.com/OPNbwhu

After: https://imgur.com/muHR4Hs

Bug: 963183
Change-Id: I2fcb44767d8cace4bf6a91d511ee0590d2f732e1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1965851
Reviewed-by: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Amanda Baker <ambake@microsoft.com>
Reviewed-by: Lorne Mitchell <lomitch@microsoft.com>
Commit-Queue: Brandon Goddard <brgoddar@microsoft.com>