Animation Timeline keyboard focus and interactivity
The Animation tool allows users the ability to change animation duration,
start time, and keyframe length via the mouse.

This patch adds keyboard interactivity to the scenarios above. Due to the
dependence of the `MouseEvent` type in the `elementDragStart`,
`elementDrag`, and `elementDragEnd` callbacks passed to
`UI.installDragHandle()`, I had to create an AnimationUI specific method
to set the keyboard event listener. This also required creating a keydown
handler that is a combination of the `_mouseDown`, `_mouseMove`, and
`_mouseUp` instance methods functionality sans calls to `MouseEvent`
specific properties.

One such property was `clientX`, which was used to determine the next
draw location on the timeline for an element being dragged. For
keyboard interactivity a new instance variable
`_keyboardMovementRateMs` was created to set the movement rate.
This is currently a constant set to 50ms, but we may want to listen to
feedback to see if we need to create a variable rate based on animation
properties.

GIF: https://i.imgur.com/RlEQEga.gif
Bug: 963183
Change-Id: Ib5c94162a582a06729cf71f0afd965b0c2465347
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1920984
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Commit-Queue: Kalon Hinds <kahinds@microsoft.com>