Rolled MediaTable into EventDisplayTable
Also cleaned up some the table elements to get json views - this
required a small change to the constructor for the JSONView class
which allows the collapsible viewer to start collapsed instead of
expanded. Since it's an optional arg, it shouldn't doesn't affect
any other uses of the class.

Here are some screenshots:
https://files.tedm.io/dtss/ev/1.png
https://files.tedm.io/dtss/ev/2.png

Bug: 794255
Change-Id: Ib83636a358e5eeeb47f958fd8046e80b0896a424
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/1996026
Reviewed-by: Jan Scheffler <janscheffler@chromium.org>
Commit-Queue: Ted Meyer <tmathmeyer@chromium.org>