Don't set markers for multi-line breakpoints
Multi-line breakpoints invoke quirky UI behavior currently. When a
breakpoint is set that (e.g., through source mapping) affects multiple
lines, inline breakpoint decorations are added on all lines.
Decorations are however tracked by editor line number, which means
when the same breakpoint is added multiple times, corresponding
decoriations in different lines are not updated, which leads to
markers stacking on the other lines.

This CL is a quick workaround that only adds inline decorations on the
line that was clicked to set the breakpoint.

Bug: chromium:1043614
Change-Id: I45dd368e7a11e40da98aa7315c077c6a5cd6187e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2013143
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>