Standardize license headers
Many of the license headers were incorrect or corrupted. To prepare for
a check that verifies correct license headers are added to source code,
we should fix up the existing license headers to use the correct format.

The license headers that include non-Google/non-Chromium authors are
left as-is.

Change-Id: Iab9bdc41d815ac7ab07ccacf44818f3b56a75230
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2016745
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>