Ensure pseudo-elements are not exposed through $0 etc.
Pseudo-elements such as those created through ::before and ::after
in CSS are not accessible as DOM nodes, per the DOM Standard.

Previously, it was possible to get access to them in the DevTools
Console by using `$0` (most recently inspected element), `$1`, and
so on after inspecting the pseudo-element in the Elements panel.
Whoops!

This patch prevents exposing access to pseudo-elements directly
through `$0` etc. similarly to how this is avoided for shadow roots.

To verify the fix, you can use the following minimal test case:

    data:text/html,<style>p::before{content:'X'}</style><p>

Inspect the `::before` pseudo-element in DevTools, enter `$0` in the
Console, and verify it refers to the perfectly ordinary `<p>` element
and not some weird internal wrapper for the `::before` pseudo-element.

Fixed: chromium:1045857
Change-Id: I766d48cdb90e4313ed98d406ebf5ac67511a2032
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2027428
Auto-Submit: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>