[e2e] Update helper to use hosted mode instead of file protocol
We can use the hosted mode server booted as part of the e2e run
to serve files instead of going via the file protocol. This means
that content that needs to be served as http(s) will work.
This CL updates the helper path to use the hosted mode path
by default.

Bug: 1047667
Change-Id: I10767df85b8205dddc81fd7d40e497511cbfdc33
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2031118
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>